Refactor useGeocoding hook
