+++ backend/data/data.js
@@ -4,7 +4,7 @@ export const data = [
     userId: '115068485855702141305',
     type: 'device',
     name: 'Device 1',
-    description: 'Device 1 description, hehe',
+    description: 'Device 1 description',
     gatewayID: 'g3',
     location: {
       lat: 44.8167,
@@ -19,81 +19,19 @@ export const data = [
   {
     id: '2',
     userId: '115068485855702141305',
-    type: 'device',
-    name: 'Device 1',
-    description: 'Device 1 description, hehe',
-    gatewayID: 'g3',
+    type: 'gateway',
+    name: 'Gateway 1',
+    key: 'key1234',
     location: {
-      lat: 44.8167,
-      lng: 20.4661,
-    },
-    metrics: {
-      pm2_5: 7.8,
-      co2: 695.5,
-      temp: 13,
+      lat: 44.8189,
+      lng: 20.4645,
     },
   },
   {
     id: '3',
     userId: '115068485855702141305',
-    type: 'gateway',
-    name: 'Gateway 1',
-    key: 'key1234',
-    location: {
-      lat: 44.8167,
-      lng: 20.4661,
-    },
-  },
-  {
-    id: '4',
-    userId: '115068485855702141305',
     type: 'device',
     name: 'Device 2',
-    description: 'Outdoor sensor in Zemun',
-    gatewayID: 'g1',
-    location: {
-      lat: 44.851,
-      lng: 20.3983,
-    },
-    metrics: {
-      pm2_5: 15.2,
-      co2: 720.3,
-      temp: 19.4,
-    },
-  },
-  {
-    id: '5',
-    userId: '115068485855702141305',
-    type: 'gateway',
-    name: 'Gateway 2',
-    key: 'zemun456',
-    location: {
-      lat: 44.85,
-      lng: 20.4,
-    },
-  },
-  {
-    id: '6',
-    userId: '115068485855702141305',
-    type: 'device',
-    name: 'Device 3',
-    description: 'Sensor in New Belgrade',
-    gatewayID: 'g2',
-    location: {
-      lat: 44.8074,
-      lng: 20.385,
-    },
-    metrics: {
-      pm2_5: 10.5,
-      co2: 660.0,
-      temp: 21.2,
-    },
-  },
-  {
-    id: '7',
-    userId: 'xxxxxxx',
-    type: 'device',
-    name: 'Device 4',
     description: 'Kalemegdan air monitor',
     gatewayID: 'g1',
     location: {
